Display hearing request type column on assign hearings table (#15312)
Follows #15294
Resolves #15059

⚠️ This can only be deployed after #15294 has been merged and deployed to production

### Description
- create a new column `Hearing Type` on assign hearings table
- rename `Types` column to `Appeal Stream Type`
- break out tests from `spec/feature/hearings/schedule_veteran/build_hearsched_spec.rb` to a new test file `spec/feature/hearings/assign_hearings_table_spec.rb` and write some new tests
- fix the docket line display

### Testing Plan
Console
- run `UpdateCachedAppealsAttributesJob.new.perform_now`

Steps
- login as `BVASYELLOW` and go to `http://localhost:3000/hearings/schedule/assign`
- choose a RO which has veterans waiting
- look at `Legacy Veterans Waiting` and `AMA Veterans Waiting`

### User Facing Changes
 - [x] Screenshots of UI changes added to PR & Original Issue

 BEFORE|AFTER
 ---|---
<img width="996" alt="Screen Shot 2020-09-24 at 4 26 50 PM" src="https://user-images.githubusercontent.com/20735998/94196714-c73ba580-fe82-11ea-96e9-17cfab11e9fc.png"> |<img width="993" alt="Screen Shot 2020-09-24 at 4 25 50 PM" src="https://user-images.githubusercontent.com/20735998/94196755-d4589480-fe82-11ea-88eb-ffdd0f3f2b61.png"> |
