+++ app/models/queue_tabs/assign_hearing_tab.rb
@@ -80,7 +80,8 @@ class AssignHearingTab
       Constants.QUEUE_CONFIG.COLUMNS.CASE_DETAILS_LINK.name,
       Constants.QUEUE_CONFIG.COLUMNS.APPEAL_TYPE.name,
       Constants.QUEUE_CONFIG.POWER_OF_ATTORNEY_COLUMN_NAME,
-      Constants.QUEUE_CONFIG.SUGGESTED_HEARING_LOCATION_COLUMN_NAME
+      Constants.QUEUE_CONFIG.SUGGESTED_HEARING_LOCATION_COLUMN_NAME,
+      Constants.QUEUE_CONFIG.HEARING_REQUEST_TYPE_COLUMN_NAME
     ]
   end
