+++ client/app/queue/utils.js
@@ -118,11 +118,14 @@ const taskAttributesFromRawTask = (task) => {
     hideFromTaskSnapshot: task.attributes.hide_from_task_snapshot,
     hideFromCaseTimeline: task.attributes.hide_from_case_timeline,
     availableHearingLocations: task.attributes.available_hearing_locations,
-    // `powerOfAttorneyName` and `suggestedHearingLocation` are only present for
+    // `powerOfAttorneyName`, `suggestedHearingLocation`,
+    // `hearingRequestType`, and `isFormerTravel` are only present for
     // /hearings/scheduled/assign page, and are not returned from the API when
     // requesting the full task.
     powerOfAttorneyName: task.attributes.power_of_attorney_name,
-    suggestedHearingLocation: task.attributes.suggested_hearing_location
+    suggestedHearingLocation: task.attributes.suggested_hearing_location,
+    hearingRequestType: task.attributes.hearing_request_type,
+    isFormerTravel: task.attributes.former_travel
   };
 };
