+++ client/constants/QUEUE_CONFIG.json
@@ -11,6 +11,7 @@
   "LEGACY_ASSIGN_HEARINGS_TAB_NAME": "legacyAssignHearingsTab",
   "SUGGESTED_HEARING_LOCATION_COLUMN_NAME": "suggestedLocation",
   "POWER_OF_ATTORNEY_COLUMN_NAME": "powerOfAttorneyName",
+  "HEARING_REQUEST_TYPE_COLUMN_NAME": "hearingRequestType",
 
   "COLUMNS": {
     "APPEAL_TYPE": {
